+++ skimage/filter/rank/core_cy.pxd
@@ -15,14 +15,14 @@ cdef dtype_t _max(dtype_t a, dtype_t b)
 cdef dtype_t _min(dtype_t a, dtype_t b)
 
 
-cdef void _core(float kernel(Py_ssize_t*, float, dtype_t,
-                             Py_ssize_t, Py_ssize_t, float,
-                             float, Py_ssize_t, Py_ssize_t),
+cdef void _core(double kernel(Py_ssize_t*, double, dtype_t,
+                              Py_ssize_t, Py_ssize_t, double,
+                              double, Py_ssize_t, Py_ssize_t),
                 dtype_t[:, ::1] image,
                 char[:, ::1] selem,
                 char[:, ::1] mask,
                 dtype_t_out[:, ::1] out,
                 char shift_x, char shift_y,
-                float p0, float p1,
+                double p0, double p1,
                 Py_ssize_t s0, Py_ssize_t s1,
                 Py_ssize_t max_bin) except *
